Refactor + fix scrollview in scrollview
Now, the RecyclerView is inside a fixed, non-scrollable LinearLayout.
The exterior margins have been removed.
