+++ app/src/main/res/layout-sw840dp-land/fragment_main.xml
@@ -15,7 +15,7 @@
         android:orientation="vertical">
 
         <Button
-            android:id="@+id/random"
+            android:id="@+id/button_random"
             style="@style/Widget.MaterialComponents.Button.Icon"
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
